% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accTest.R
\name{accuracy}
\alias{accuracy}
\title{samples two rows in a matrix and check if the samples are similar or different
based on their cluster labels}
\usage{
accuracy(mat, mono_clust, bi_clust, nrun = 10000)
}
\arguments{
\item{mat}{matrix to be sampled from}

\item{mono_clust}{monoculture cluster label}

\item{bi_clust}{biculture cluster label}

\item{nrun}{number of times the resampling should be carried out.
Defaults to 10000}
}
\value{
a vector of integer values
}
\description{
This function
}
\examples{
x <- matrix(NA, nrow = 100, ncol = 3)
xx <- apply(x, 2, rnorm, 100)
xx <- cbind(xx, Mono = rep(1:2, each = 50),
            Bi = rep(1:2, times = 50))
accuracy(xx, "Mono", "Bi", nrun = 5000)
}
