% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs4_c.R
\name{bs4_c}
\alias{bs4_c}
\title{function to gate out bs4 from a debris filtered flowframe using a control flowframe.
Called within the cell_debris function.}
\usage{
bs4_c(bs4bs5, p1, p2, day1_2, others)
}
\arguments{
\item{bs4bs5}{debris filtered flowframe}

\item{p1}{same as cell_debris}

\item{p2}{same as cell_debris}

\item{day1_2}{debris gated mono_control}

\item{others}{position of not yet identified particles}
}
\value{
\itemize{
        \item \strong{bs4_reduced -} reduced flowframe containing BS4
        \item \strong{others_nk -} position of unknown particles
        \item \strong{bs4_others -} position of BS4s
        \item \strong{others_bs5 -} position of BS5s, if present
 }
}
\description{
function to gate out bs4 from a debris filtered flowframe using a control flowframe.
Called within the cell_debris function.
}
