% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phycocyanin.R
\name{bs4_nc}
\alias{bs4_nc}
\title{gates out or assign indicators to phytoplankton cells based on the expression of
Phycocyanin.}
\usage{
bs4_nc(bs4bs5, ch1, ch2, others, ph = 0.1, to_retain = c("refined",
  "potential"))
}
\arguments{
\item{bs4bs5}{flowframe with debris (left) removed.}

\item{ch1}{first flowcytometer channel measuring first pigmentation.
For example, channel measuring phycocyanin}

\item{ch2}{second flowcytometer channel measuring second pigmentation.
For example, channel measuring phycoerytherin}

\item{others}{row numbers for non-debris events. This is provided by the
\code{\link{debris_nc}} or \code{\link{debris_inc}} function.}

\item{to_retain}{should potential candidates be retained or further gating be applied
to filter out only certain phytoplankton cells.}
}
\value{
list containing; \itemize{
\item \strong{syn_reduced -} flowframe containing only phytoplankton cells
\item \strong{others_nk -} unidentified particle positions
\item \strong{syn_pos -} phytoplankton cells positions
\item \strong{others_nk2 -} other unidentified particle positions
}
}
\description{
This function takes in a flowframe with debris removed and identifies
             phytoplankton cell population in the provided frame.
}
\details{
The function uses the \code{\link[flowDensity]{getPeaks}} and
         \code{\link[flowDensity]{deGate}} functions in the \emph{flowDensity} package to
         identify peaks and identify cut-off points between these peaks. This function is
         not designed to be called in isolation, if called in isolation an error will be
         returned. It is preferably called on the results from \code{\link{debris_nc}} or
         \code{\link{debris_inc}} function. A graph with horizontal
         and vertical lines used in separating the populations is returned and
         if \emph{to_retain = "refined"}, a circle made of dashed lines is drawn around
         phytoplankton cell population points.
}
\seealso{
\code{\link{bs5_nc}}
}
