% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs5_c.R
\name{bs5_c}
\alias{bs5_c}
\title{function to gate out bs5 from a debris filtered flowframe using a control flowframe.
Called within the cell_debris function.}
\usage{
bs5_c(bs4bs5, p1, p2, day1_2, others)
}
\arguments{
\item{bs4bs5}{debris filtered flowframe}

\item{p1}{same as cell_debris}

\item{p2}{same as cell_debris}

\item{day1_2}{debris gated mono_control}

\item{others}{position of not yet identified particles}
}
\value{
\itemize{
        \item \strong{bs5_reduced -} reduced flowframe containing BS5
        \item \strong{others_nk -} position of unknown particles
        \item \strong{others_bs52 -} position of BS5s
        \item \strong{others_bs4 -} position of BS4s, if present
}
}
\description{
function to gate out bs5 from a debris filtered flowframe using a control flowframe.
Called within the cell_debris function.
}
