% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{cluster_assign}
\alias{cluster_assign}
\title{function to assign particles to cluster based on a supplied classifier. It is useful for the EM clustering approach.}
\usage{
cluster_assign(flowfile, classifier)
}
\arguments{
\item{flowfile}{flowframe to be plotted}

\item{classifier}{minimum desired probability}
}
\description{
function to assign particles to cluster based on a supplied classifier. It is useful for the EM clustering approach.
}
