% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvt_approach.R
\name{cyano_emclustering}
\alias{cyano_emclustering}
\title{clusters BS4, BS5 and Debris in a flowfile using an EM style algorithm.
Data is clustered into 5 clusters (automatically reduces this number if need be)}
\usage{
cyano_emclustering(flowfile, channels, mu = NULL, sigma = NULL,
  ncluster = 5, min.itera = 20)
}
\arguments{
\item{flowfile}{flowframe to be clustered.}

\item{channels}{channels to use for the clustering}

\item{mu}{pre-specified mean matrix for the clusters. Number of rows should equal ncluster and number
of columns should equal length(channels). Defaults to NULL and can be computed from the data internally.}

\item{sigma}{pre-specified list of variance-covariace matrix for the clusters. Each element of the list should contain a square matrix
of variance-covariance matrix with length equal ncluster.}

\item{ncluster}{number of cluster to desired}
}
\value{
list containing; \itemize{
\item \strong{percentages -}
\item \strong{mus -}
\item \strong{sigmas -}
\item \strong{result -}
}
}
\description{
clusters BS4, BS5 and Debris in a flowfile using an EM style algorithm.
Data is clustered into 5 clusters (automatically reduces this number if need be)
}
\examples{
\dontrun{
cyano_emclustering(flowfile = flowfile, channel1 = c("RED.B.HLin", "YEL.B.HLin", "FSC.HLin", "RED.R.HLin"),
ncluster = 5, min.itera = 20)
}


}
