% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyanoFilter.R
\docType{package}
\name{cyanoFilter}
\alias{cyanoFilter}
<<<<<<< HEAD
\title{cyanoFilter: A package to identify and cluster phytoplankton cells contained in flow cytometry data.}
=======
\alias{cyanoFilter-package}
\title{cyanoFilter: A package to identify and/or assign indicators to BS4, BS5 cyanobacteria cells contained in water sample.}
>>>>>>> 74ebbc1f0b57e1f0410570dfc73e0a04247a1d3a
\description{
The package provides two categories of functions:
\emph{metafile} preprocessing functions and \emph{fcsfile} processing functions.
}
\section{metafile preprocessing functions}{

<<<<<<< HEAD
          This set of functions (\code{\link{goodfcs}} and \code{\link{retain}}) helps to
          identify the appropriate fcs file to read.
=======
          This set of functions (\code{\link{goodfcs}} and \code{\link{retain}}) helps to identify the appropriate fcs file to read.
>>>>>>> 74ebbc1f0b57e1f0410570dfc73e0a04247a1d3a
}

\section{fcsfile processing functions}{

<<<<<<< HEAD
         These functions (\code{\link{nona}} and \code{\link{noneg}}, \code{\link{noneg}},
         \code{\link{phyto_filter}}, \code{\link{celldebris_emclustering}})
         works on the fcs file to identify the phytoplankton populations contained in
         the fcs file.
=======
         These functions (\code{\link{nona}} and \code{\link{noneg}}, \code{\link{noneg}}, \code{\link{celldebris_nc}}, \code{\link{celldebris_emclustering}})
         works on the fcs file to identify the cell populations contained in the sample that generated this file.
>>>>>>> 74ebbc1f0b57e1f0410570dfc73e0a04247a1d3a
}

