% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debris_c.R
\name{debris_c}
\alias{debris_c}
\title{function to gate out bs4 from a debris filtered flowframe using a control flowframe.
Called within the cell_debris function.}
\usage{
debris_c(flowframe, p1, p2, control_flowFrame)
}
\arguments{
\item{flowframe}{supplied flowframe}

\item{p1}{same as cell_debris}

\item{p2}{same as cell_debris}

\item{control_flowFrame}{control flowframe to used to identify debris}
}
\value{
list containing;
    \itemize{
        \item \strong{bs4bs5 -} debris filtered flowframe
        \item \strong{others -} position of yet to be identified particles
        \item \strong{debs -} position of debris
        }
}
\description{
function to gate out bs4 from a debris filtered flowframe using a control flowframe.
Called within the cell_debris function.
}
