% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debris_invasion_nc.R
\name{debris_inc}
\alias{debris_inc}
\title{gates out or assign indicators to debris contained a flowframe.}
\usage{
debris_inc(flowframe, p1, p2)
}
\arguments{
\item{flowframe}{flowframe with.}

\item{p1}{first flowcytometer channel that can be used to separate BS4 cells from the rest, e.g. "RED.B.HLin".}

\item{p2}{second flowcytometer channel that can be used to separate BS4 cells from the rest, e.g. "YEL.B.HLin"}
}
\value{
list containing; \itemize{
\item \strong{bs4bs5 -}
\item \strong{deb_pos -}
\item \strong{bs4bs5_pos -}
}
}
\description{
gates out or assign indicators to debris contained a flowframe.
}
\examples{
\dontrun{
debris_inc(flowframe = flowfile, p1 = "RED.B.HLin", p2 = "YEL.B.HLin")
}

}
