% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phytoFilterClass.R
\name{fullFlowframe,PhytoFilter-method}
\alias{fullFlowframe,PhytoFilter-method}
\title{accesor method for reduced flowframe(PhytoFilter class)}
\usage{
\S4method{fullFlowframe}{PhytoFilter}(x)
}
\arguments{
\item{x}{an object of class PhytoFilter}
}
\value{
fullFlowframe method for PhytoFilter
}
\description{
accesor method for reduced flowframe(PhytoFilter class)
}
