% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phytoFilterClass.R
\name{fullFlowframe}
\alias{fullFlowframe}
\title{generic function for extracting the full flowframe}
\usage{
fullFlowframe(x)
}
\arguments{
\item{x}{an object of either class PhytoFilter, MarginEvents or DebrisFilter}
}
\value{
generic to extract fullFlowframe
}
\description{
generic function for extracting the full flowframe
}
