% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getChannels.R
\name{get_channel}
\alias{get_channel}
\title{returns the channel with more than one peak present. It returns NA if there is only one peak present.}
\usage{
get_channel(flowfile, ch, ph)
}
\arguments{
\item{flowfile}{flowframe after debris are removed.}

\item{ch}{channel to be checked for multiple peaks.}

\item{ph}{maximum peak height to be ignored. This allows ignoring of tiny peaks that could
affect the gating process.}
}
\description{
returns the channel with more than one peak present. It returns NA if there is only one peak present.
}
