% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{ggpairsDens}
\alias{ggpairsDens}
\title{produces a scatter plot of the expression matrix of the flowframe. If a cluster variable is given,
it assigns different colors to the clusters.}
\usage{
ggpairsDens(flowfile, channels = NULL, group = NULL, notToPlot = NULL, ...)
}
\arguments{
\item{flowfile}{flowframe to be plotted}

\item{channels}{a character vector of length 2 or more. It must contain channel names in the flowfile.}

\item{group}{cluster groups. It must be equal to the number of particles in the flowfile. If group is
null cluster boundaries are not drawn.}

\item{notToPlot}{columns not to plot. This is especially useful for for plotting all columns in a}

\item{...}{not used at the moment}
}
\value{
a ggplot object
}
\description{
produces a scatter plot of the expression matrix of the flowframe. If a cluster variable is given,
it assigns different colors to the clusters.
}
\examples{
\donttest{

 # example without clustering
flowfile_path <- system.file("extdata", "B4_18_1.fcs", package = "cyanoFilter",
              mustWork = TRUE)
flowfile <- flowCore::read.FCS(flowfile_path, alter.names = TRUE,
                               transformation = FALSE, emptyValue = FALSE,
                               dataset = 1) #FCS file contains only one data object
flowfile_nona <- cyanoFilter::nona(x = flowfile)
flowfile_noneg <- cyanoFilter::noneg(x = flowfile_nona)
flowfile_logtrans <- lnTrans(x = flowfile_noneg, c('SSC.W', 'TIME'))
ggpairsDens(flowfile = flowfile_logtrans,
            channels = c("FSC.HLin", "RED.R.HLin", "RED.B.HLin", "NIR.R.HLin"))

 # example plot after clustering
flowfile_path <- system.file("extdata", "B4_18_1.fcs", package = "cyanoFilter",
                             mustWork = TRUE)
flowfile <- flowCore::read.FCS(flowfile_path, alter.names = TRUE,
                               transformation = FALSE, emptyValue = FALSE,
                               dataset = 1) #FCS file contains only one data object
flowfile_nona <- cyanoFilter::nona(x = flowfile)
flowfile_noneg <- cyanoFilter::noneg(x = flowfile_nona)
flowfile_logtrans <- cyanoFilter::lnTrans(x = flowfile_noneg, c('SSC.W', 'TIME'))
cells_nonmargin <- cellmargin(flowframe = flowfile_logtrans, Channel = 'SSC.W',
           type = 'estimate', y_toplot = "FSC.HLin")
cells_nodebris <- debris_nc(flowframe = cells_nonmargin$reducedflowframe,
                            ch_chlorophyll = "RED.B.HLin",
                            ch_p2 = "YEL.B.HLin",
                            ph = 0.05)
fin <- phyto_filter(flowfile = cells_nodebris$syn,
              pig_channels = c("RED.B.HLin", "YEL.B.HLin", "RED.R.HLin"),
              com_channels = c("FSC.HLin", "SSC.HLin"))

ggpairsDens(flowfile = fin$flowframe_proportion,
                    channels = c("RED.B.HLin", "YEL.B.HLin", "RED.R.HLin", "FSC.HLin"),
                    cluster_var = "Clusters")



 }

}
