% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phytoFilterClass.R
\name{is.cyanoFilter}
\alias{is.cyanoFilter}
\title{function to check if object is of class cyanoFilter}
\usage{
is.cyanoFilter(x)
}
\arguments{
\item{x}{any R object}
}
\description{
function to check if object is of class cyanoFilter
}
