% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phytoFilterClass.R
\name{is.DebrisFilter}
\alias{is.DebrisFilter}
\title{function to check if object is of class cyanoFilter(DebrisFilter)}
\usage{
is.DebrisFilter(x)
}
\arguments{
\item{x}{any R object}
}
\value{
TRUE if object is of class DebrisFilter. FALSE otherwise
}
\description{
function to check if object is of class cyanoFilter(DebrisFilter)
}
\examples{
 x <- c(1, 5, 4)
 is.DebrisFilter(x)
}
