% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phytoFilterClass.R
\name{is.flowFrame}
\alias{is.flowFrame}
\title{function to check if object is a flowFrame}
\usage{
is.flowFrame(x)
}
\arguments{
\item{x}{any R object}
}
\value{
TRUE if object is a flowFrame. FALSE otherwise
}
\description{
function to check if object is a flowFrame
}
\examples{
 x <- c(1, 5, 4)
 is.flowFrame(x)
}
