% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phytoFilterClass.R
\name{is.flowSet}
\alias{is.flowSet}
\title{function to check if object is a flowSet}
\usage{
is.flowSet(x)
}
\arguments{
\item{x}{any R object}
}
\value{
TRUE if object is a flowSet. FALSE otherwise
}
\description{
function to check if object is a flowSet
}
\examples{
 x <- c(1, 5, 4)
 is.flowSet(x)
}
