% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phytoFilterClass.R
\name{is.MarginEvents}
\alias{is.MarginEvents}
\title{function to check if object is of class cyanoFilter(MarginEvents)}
\usage{
is.MarginEvents(x)
}
\arguments{
\item{x}{any R object}
}
\value{
TRUE if object is of class MarginEvents. FALSE otherwise
}
\description{
function to check if object is of class cyanoFilter(MarginEvents)
}
\examples{
 x <- c(1, 5, 4)
 is.MarginEvents(x)
}
