% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phytoFilterClass.R
\name{is.PhytopFilter}
\alias{is.PhytopFilter}
\title{function to check if object is of class cyanoFilter(PhytoFilter)}
\usage{
is.PhytopFilter(x)
}
\arguments{
\item{x}{any R object}
}
\value{
TRUE if object is of class PhytoFilter. FALSE otherwise
}
\description{
function to check if object is of class cyanoFilter(PhytoFilter)
}
\examples{
 x <- c(1, 5, 4)
 is.PhytopFilter(x)

}
