% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lnTrans.R
\name{lnTrans}
\alias{lnTrans}
\title{log transforms the expression matrix of a flowframe}
\usage{
lnTrans(x, notToTransform = c("SSC.W", "TIME"))
}
\arguments{
\item{x}{flowframe to be transformed}

\item{notToTransform}{columns not to be transformed}
}
\value{
\strong{flowframe} with log transformed expression matrix
}
\description{
log transforms the expression matrix of a flowframe
}
\examples{
\dontrun{
lnTrans(x = flowfile, c('YEL.B.HLin', 'TIME'))
}

}
