% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvnorm.R
\name{mvnorm}
\alias{mvnorm}
\title{multivariate normal density}
\usage{
mvnorm(x, mu, sigma)
}
\arguments{
\item{x}{matrix to compute density on}

\item{mu}{mean vector}

\item{sigma}{variance covariance matrix}
}
\value{
vector of density
}
\description{
multivariate normal density
}
