% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newFlowframe.R
\name{new_flowframe}
\alias{new_flowframe}
\title{takes a flowframe, a group indicator and formulates another flowframe with group indicator as part of the
expression matrix of the new flowframe.}
\usage{
new_flowframe(flowfile, group, togate)
}
\arguments{
\item{flowfile}{flowframe after debris are removed.}

\item{group}{cluster group to be added to the expression matrix}

\item{togate}{channel detected to have more than one peak based on \code{\link{get_channel}}}
}
\description{
takes a flowframe, a group indicator and formulates another flowframe with group indicator as part of the
expression matrix of the new flowframe.
}
