% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pigment_gate.R
\name{pigment_gate}
\alias{pigment_gate}
\title{gates out or assign indicators to phytoplankton cells based on the expression of
the measured pigments.}
\usage{
pigment_gate(flowfile, pig_channels, ph = 0.05,
  width_channel = "SSC.W")
}
\arguments{
\item{flowfile}{flowframe after debris are removed.}

\item{pig_channels}{flowcytometer channels measuring phytoplankton pigmentations.}

\item{ph}{maximum peak height to be ignored. This allows ignoring of tiny peaks that could
affect the gating process.}

\item{width_channel}{a second channel with only one peak present.}
}
\value{
list containing; \itemize{
\item \strong{full_flowframe -} flowframe containing only phytoplankton cells
\item \strong{phy_pos_nk -} unidentified particle positions
\item \strong{syn_pos -} phytoplankton cells positions
\item \strong{phy_pos_nk2 -} other unidentified particle positions
}
}
\description{
This function takes in a flowframe with debris removed and identifies
             phytoplankton cell population in the provided frame.
}
\details{
The function uses the \code{\link[flowDensity]{getPeaks}} and
         \code{\link[flowDensity]{deGate}} functions in the \emph{flowDensity} package to
         identify peaks and identify cut-off points between these peaks. This function is
         not designed to be called in isolation, if called in isolation an error will be
         returned. It is preferably called on the results from \code{\link{debris_nc}} or
         \code{\link{debris_inc}} function. A graph with horizontal
         and vertical lines used in separating the populations is returned and
         if \emph{to_retain = "refined"}, a circle made of dashed lines is drawn around
         phytoplankton cell population points.
}
\seealso{

}
