% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phytoFilterClass.R
\name{plot,DebrisFilter,ANY-method}
\alias{plot,DebrisFilter,ANY-method}
\title{plot method for DebrisFilter objects}
\usage{
\S4method{plot}{DebrisFilter,ANY}(x)
}
\arguments{
\item{x}{an object of class DebrisFilter}
}
\value{
object of class ggplot
}
\description{
plot method for DebrisFilter objects
}
