% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phytoFilterClass.R
\name{plot,MarginEvents,ANY-method}
\alias{plot,MarginEvents,ANY-method}
\title{plot method for MarginEvents objects}
\usage{
\S4method{plot}{MarginEvents,ANY}(x)
}
\arguments{
\item{x}{an object of class MarginEvents}
}
\value{
object of class ggplot
}
\description{
plot method for MarginEvents objects
}
