% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phytoFilterClass.R
\name{plot,PhytopFilter,ANY-method}
\alias{plot,PhytopFilter,ANY-method}
\title{plot method for PhytoFilter objects}
\usage{
\S4method{plot}{PhytopFilter,ANY}(x)
}
\arguments{
\item{x}{an object of class PhytoFilter}
}
\value{
object of calss ggplot
}
\description{
plot method for PhytoFilter objects
}
