% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot}
\alias{plot}
\title{plot method for cyanoFilter objects}
\usage{
plot(x, ch = NULL, ...)
}
\arguments{
\item{x}{object of class cyanoFilter to plotted}

\item{ch}{channels to be plotted. This must be supplied if x is a flowFrame.}

\item{...}{other arguments. Not used at the moment}
}
\description{
plot method for cyanoFilter objects
}
