% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phytoFilterClass.R
\name{reducedFlowframe,DebrisFilter-method}
\alias{reducedFlowframe,DebrisFilter-method}
\title{accesor method for reduced flowframe (DebrisFilter class)}
\usage{
\S4method{reducedFlowframe}{DebrisFilter}(x)
}
\arguments{
\item{x}{an object of class DebrisFilter}
}
\value{
reduced flowFrame method for DebrisFilter
}
\description{
accesor method for reduced flowframe (DebrisFilter class)
}
