% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phytoFilterClass.R
\name{reducedFlowframe,PhytoFilter-method}
\alias{reducedFlowframe,PhytoFilter-method}
\title{accesor method for reduced flowframe(PhytoFilter class)}
\usage{
\S4method{reducedFlowframe}{PhytoFilter}(x)
}
\arguments{
\item{x}{an object of class PhytoFilter}
}
\value{
reduced flowFrame method for PhytoFilter
}
\description{
accesor method for reduced flowframe(PhytoFilter class)
}
