% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phytoFilterClass.R
\name{reducedFlowframe}
\alias{reducedFlowframe}
\title{generic function for extracting the reduced flowframe}
\usage{
reducedFlowframe(x)
}
\arguments{
\item{x}{an object of either class PhytoFilter, MarginEvents or DebrisFilter}
}
\value{
generic for reduced flowFrame
}
\description{
generic function for extracting the reduced flowframe
}
