% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refine_gate.R
\name{refine_gate}
\alias{refine_gate}
\title{returns the position of the identified clusters after 2D gating with the size channel}
\usage{
refine_gate(flowfile, phy_pos, other_channel, togate, percent = 0.6)
}
\arguments{
\item{flowfile}{flowframe after debris are removed.}

\item{phy_pos}{row numbers for events belonging to each cluster. This is provided by the
\code{\link{row_numbers}} function.}

\item{other_channel}{extracted as the width_channel in \code{\link{oneDgate}}}

\item{togate}{channels detected to have more than one peak present. Provide by the
\code{\link{get_channel}} function.}

\item{percent}{cut-off percentile for other_channel.}
}
\description{
returns the position of the identified clusters after 2D gating with the size channel
}
