% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retain.R
\name{retain}
\alias{retain}
\title{Decides if a file should be retiained or removed based on its status.}
\usage{
retain(meta_files, make_decision = c("maxi", "mini"),
  Status = "Status", CellspML = "CellspML")
}
\arguments{
\item{meta_files}{dataframe from meta file that has been preprocessed by the \code{\link{goodfcs}} function.}

\item{make_decision}{decision to be made should more than one \eqn{cells/\mu L} be good.}

\item{Status}{column name in meta_files containing status obtained from the \code{\link{goodfcs}} function.}

\item{CellspML}{column name in meta_files containing \eqn{cells/\mu L} measurements.}
}
\value{
a character vector with entries "Retain" for a file to be retained or "No!" for a file to be discarded.
}
\description{
Function to determine what files to retain and finally read from the flow cytometer FCS file.
}
\details{
It is typically not known in advance which dilution level would result in the desired \eqn{cells/\mu L}, therefore
         the samples are ran through the flow cytometer at two or more dilution levels. Out of these, one has to decide which
         to retain and finally use for further analysis. This function and \code{\link{goodfcs}} are to help you decide that.
         If more than one of the dilution levels are judged good, the option \emph{make_decision = "maxi"} will give "Retain" to the
         row with the maximum \eqn{cells/\mu L} while the opposite occurs for \emph{make_decision = "mini"}.
}
\examples{
\dontrun{
retain(meta_files = dataframe, make_decision = "maxi")
}

}
\seealso{
\code{\link{goodfcs}}
}
