% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phytoFilterClass.R
\name{summary,MarginEvents-method}
\alias{summary,MarginEvents-method}
\title{takes a flowframes, a vector of channels, cluster indicator and return 
desired summaries per cluster}
\usage{
\S4method{summary}{MarginEvents}(object, channels = NULL, ...)
}
\arguments{
\item{object}{An object of class MarginEvents to be summarised.}

\item{channels}{channels whose summaries are to be computed}

\item{...}{other arguments. Not used at the moment}
}
\value{
list containing the required summaries
}
\description{
takes a flowframes, a vector of channels, cluster indicator and return 
desired summaries per cluster
}
\examples{
flowfile_path <- system.file("extdata", "B4_18_1.fcs", 
              package = "cyanoFilter",
              mustWork = TRUE)
flowfile <- flowCore::read.FCS(flowfile_path, alter.names = TRUE,
                               transformation = FALSE, emptyValue = FALSE,
                               dataset = 1) 
flowfile_nona <- cyanoFilter::nona(x = flowfile)
flowfile_noneg <- cyanoFilter::noneg(x = flowfile_nona)
flowfile_logtrans <- cyanoFilter::lnTrans(x = flowfile_noneg, 
c('SSC.W', 'TIME'))
cells_nonmargin <- cellmargin(flowframe = flowfile_logtrans, 
Channel = 'SSC.W',
           type = 'estimate', y_toplot = "FSC.HLin")
summary(object = reducedFlowframe(cells_nonmargin),
        channels = c("RED.B.HLin", "YEL.B.HLin", "RED.R.HLin"))

}
