% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs4_nc.R
\name{bs4_nc}
\alias{bs4_nc}
\title{gates out or assign indicators to BS4 cyano cells from a flowframe.}
\usage{
bs4_nc(bs4bs5, p1, p2, others, to_retain = c("refined", "potential"))
}
\arguments{
\item{bs4bs5}{flowframe with debris gated.}

\item{p1}{first flowcytometer channel that can be used to separate BS4 cells from the rest, e.g. "RED.B.HLin".}

\item{p2}{second flowcytometer channel that can be used to separate BS4 cells from the rest, e.g. "YEL.B.HLin"}

\item{others}{row numbers for non-debris events. This is provided by the debris_nc or debris_inc function.}

\item{retain}{should potential candidates be retained or further gating be applied to filter out only certain BS4 cells.}
}
\value{
list containing; \itemize{
\item \strong{bs4_reduced -}
\item \strong{others_nk -}
\item \strong{bs4_pos -}
\item \strong{others_nk2 -}
}
}
\description{
gates out or assign indicators to BS4 cyano cells from a flowframe.
}
\examples{
\dontrun{
bs4_nc(bs4bs5 = flowfile, p1 = "RED.B.HLin", p1 = "YEL.B.HLin", others = b4b5_others, to_retain = "refined")
}


}
