% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Margin_Cells.R
\name{cellmargin}
\alias{cellmargin}
\title{Removes or assign indicators to margin events
1 = not margin 0=margin}
\usage{
cellmargin(flow.frame, Channel = "SSC.W", type = c("manual", "estimate"),
  cut = NULL)
}
\description{
Removes or assign indicators to margin events
1 = not margin 0=margin
}
