% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_extract.R
\name{cluster_extractp}
\alias{cluster_extractp}
\title{takes a flowframe, name of cluster column and extracts part of flowframe that makes up proportion.}
\usage{
cluster_extractp(flowfile, cluster_var = "Clusters", proportion = 0.8)
}
\arguments{
\item{flowfile}{flowframe after debris are removed.}

\item{cluster_var}{column name in expression matrix containing the cluter indicators}

\item{proportion}{value between 0 and 1 indicating percentage of the total particles wanted}
}
\description{
takes a flowframe, name of cluster column and extracts part of flowframe that makes up proportion.
}
