% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{cluster_plot}
\alias{cluster_plot}
\title{plots the expression matrix of a flowframe analysed with celldebris_emclustering.}
\usage{
cluster_plot(flowfile, channel1 = "RED.B.HLin",
  channel2 = "YEL.B.HLin", mus = NULL, tau = NULL)
}
\arguments{
\item{flowfile}{flowframe to be plotted}

\item{channel1}{column in expression matrix not to be plotted}

\item{channel2}{column in expression matrix not to be plotted}

\item{mus}{matrix of means obtained from celldebris_emclustering}

\item{tau}{vector of cluster weights obtained from celldebris_emclustering}
}
\description{
plots the expression matrix of a flowframe analysed with celldebris_emclustering.
}
