% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexity_gate.R
\name{complexity_gate}
\alias{complexity_gate}
\title{gates out or assign indicators to phytoplankton cells based on the expression of
measured cell complexity channels.}
\usage{
complexity_gate(flowfile, com_channels, ph = 0.05)
}
\arguments{
\item{flowfile}{flowframe to be processed.}

\item{com_channels}{flowcytometer channels measuring cell complexity.}

\item{ph}{maximum peak height to be ignored. This allows ignoring of tiny peaks that could
affect the gating process.}
}
\value{
list containing; \itemize{
\item \strong{full_flowframe -} flowframe containing only phytoplankton cells
\item \strong{phy_ind -} indicator of phytoplankton cells
\item \strong{gated_channels -} channels with more than two peaks found
}
}
\description{
This function takes in a flowframe with debris removed and identifies
             the different phytoplankton cell population based on cell complexity.
}
\details{
The function uses the \code{\link[flowDensity]{getPeaks}} and
         \code{\link[flowDensity]{deGate}} functions in the \emph{flowDensity} package to
         identify peaks and identify cut-off points between these peaks.
}
