% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debris_nc.R
\name{debris_nc}
\alias{debris_nc}
\title{gates out or assign indicators to BS4 cyano cells from a flowframe.}
\usage{
debris_nc(flowframe, p1, p2)
}
\arguments{
\item{flowframe}{flowframe with.}

\item{p1}{first flowcytometer channel that can be used to separate BS4 cells from the rest, e.g. "RED.B.HLin".}

\item{p2}{second flowcytometer channel that can be used to separate BS4 cells from the rest, e.g. "YEL.B.HLin"}

\item{others}{row numbers for non-debris events. This is provided by the debris_nc or debris_inc function.}

\item{retain}{should potential candidates be retained or further gating be applied to filter out only certain BS4 cells.}
}
\value{
list containing; \itemize{
\item \strong{bs4_reduced -}
\item \strong{others_nk -}
\item \strong{bs4_pos -}
\item \strong{others_nk2 -}
}
}
\description{
gates out or assign indicators to BS4 cyano cells from a flowframe.
}
\examples{
\dontrun{
debris_nc(flowframe = flowfile, p1 = "RED.B.HLin", p2 = "YEL.B.HLin")
}


}
