% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phytoFilterClass.R
\name{fullFlowframe,MarginEvents-method}
\alias{fullFlowframe,MarginEvents-method}
\title{accesor method for reduced flowframe (MarginEvent class)}
\usage{
\S4method{fullFlowframe}{MarginEvents}(x)
}
\arguments{
\item{x}{an object of class MarginEvents}
}
\value{
full Flowframe method for MarginEvents
}
\description{
accesor method for reduced flowframe (MarginEvent class)
}
