% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{ggplotDens}
\alias{ggplotDens}
\title{plots two channels of a flowframe. Note that, it takes some time to display the plot.}
\usage{
ggplotDens(flowfile, channels, ...)
}
\arguments{
\item{flowfile}{flowframe to be plotted}

\item{channels}{a character vector of length 2, must contain channel names in the flowfile.}

\item{...}{not used at the moment}
}
\value{
a ggplot object
}
\description{
plots two channels of a flowframe. Note that, it takes some time to display the plot.
}
