% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{ggplotDens2}
\alias{ggplotDens2}
<<<<<<< HEAD
\title{plots two channels of a flowframe with different colors for clusters identified.}
=======
\title{plots two channels of a flowframe. Note that, it takes some time to display the plot.}
>>>>>>> 74ebbc1f0b57e1f0410570dfc73e0a04247a1d3a
\usage{
ggplotDens2(flowfile, channels, group, ...)
}
\arguments{
\item{flowfile}{flowframe to be plotted}

\item{channels}{a character vector of length 2, must contain channel names in the flowfile.}

\item{group}{cluster groups. must be equal to the number of particles in the flow cytometer.}

\item{...}{not used at the moment}
}
\value{
a ggplot object
}
\description{
<<<<<<< HEAD
plots two channels of a flowframe with different colors for clusters identified.
=======
plots two channels of a flowframe. Note that, it takes some time to display the plot.
>>>>>>> 74ebbc1f0b57e1f0410570dfc73e0a04247a1d3a
}
