% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/good_measurement.R
\name{goodfcs}
\alias{goodfcs}
\title{indicates if measurement from a flowfile is good or bad @param fcsfile flowfile to be checked.}
\usage{
goodfcs(metafile, col_cpml = "CellspML", mxd_cellpML = 1000,
  mnd_cellpML = 50)
}
\arguments{
\item{metafile}{associated metafile to the supplied fcsfile. This is a csv file containig computed stats from the flow cytometer.}

\item{col_cpml}{column name or column number in metafile containing cell per microlitre measurements.}

\item{mxd_cellpML}{maximal accepted cell per microlitre. Flowfiles with larger cell per microlitre are termed bad.
Defaults to 1000.}

\item{mnd_cellpML}{minimum accepted cell per microlitre. Flowfiles with lesser cell per microlitre are termed bad.
Defaults to 50.}
}
\value{
character vector with length same as the number of rows in the metafile and its entries are
         \strong{good} for good files and \strong{bad} for bad files.
}
\description{
indicates if measurement from a flowfile is good or bad @param fcsfile flowfile to be checked.
}
\examples{
\dontrun{
goodfcs(metafile = flowframe, col_cpml = "CellspML", mxd_cellpML = 1000, mnd_cellpML = 50)
}
}
