% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/good_measurement.R
\name{goodfcs}
\alias{goodfcs}
\title{indicates if measurement from a flowfile is good or bad}
\usage{
goodfcs(fcsfile, metafile, CpermL, rowinmetafile, d_cellnum = 90,
  d_cellpML = 1000)
}
\description{
indicates if measurement from a flowfile is good or bad
}
