% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newFlowframe2.R
\name{new_flowframe2}
\alias{new_flowframe2}
\title{takes a list of flowframes or a flowSet with similar annotations but different expression matrices and stacks them.}
\usage{
new_flowframe2(flowframes, mode = "list")
}
\arguments{
\item{mode}{the type of object flowframe is. Can either be "list" or "set".}

\item{flowframe}{a list of flowframes or a flowSet}
}
\description{
takes a list of flowframes or a flowSet with similar annotations but different expression matrices and stacks them.
}
