% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nona.R
\name{nona}
\alias{nona}
\title{Removes NA values from the expression matrix of a flow cytometer file.}
\usage{
nona(x)
}
\arguments{
\item{x}{flowframe with expression matrix containing NAs.}
}
\value{
flowframe with expression matrix rid of NAs.
}
\description{
Removes NA values from the expression matrix of a flow cytometer file.
}
\examples{
\dontrun{
nona(x = flowfile)
}

}
