% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nona.R
\name{nona}
\alias{nona}
\title{Removes NA values from the expression matrix of a flowfile}
\usage{
nona(x)
}
\arguments{
\item{x}{is the flowframe with expression matrix containing NA}
}
\value{
a flowframe with expression matrix rid of NAs
}
\description{
Removes NA values from the expression matrix of a flowfile
}
