% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noneg.R
\name{noneg}
\alias{noneg}
\title{Removes negative values from the expression matrix}
\usage{
noneg(x)
}
\arguments{
\item{x}{is the flowframe whose expression matrix contains negative values}
}
\value{
flowframe with non-negative values in its expression matrix
}
\description{
Removes negative values from the expression matrix
}
\examples{
\dontrun{
noneg(x = flowfile)
}

}
