% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oneDgate.R
\name{oneDgate}
\alias{oneDgate}
\title{returns the labels stating the cluster of each row in a flowfile.}
\usage{
oneDgate(flowfile, togate)
}
\arguments{
\item{flowfile}{flowframe after debris are removed.}

\item{togate}{channels detected to have more than one peak present. Provide by the
\code{\link{get_channel}} function.}
}
\description{
returns the labels stating the cluster of each row in a flowfile.
}
