% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{pair_plot}
\alias{pair_plot}
\title{plots the expression matrix of a flowframe. Note that, it takes some time to display the plot.}
\usage{
pair_plot(x, notToPlot = c("TIME"))
}
\arguments{
\item{x}{flowframe to be plotted}

\item{notToPlot}{column in expression matrix not to be plotted}
}
\description{
plots the expression matrix of a flowframe. Note that, it takes some time to display the plot.
}
\examples{
\dontrun{ pair_plot(x = flowfile, notToPlot = "TIME") }
}
