% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{pairs_plot}
\alias{pairs_plot}
\title{produces a scatter plot of the expression matrix of a flowframe. Note that, it takes some time to display the plot.}
\usage{
pairs_plot(x, notToPlot = c("TIME"), ...)
}
\arguments{
\item{x}{flowframe to be plotted}

\item{notToPlot}{column in expression matrix not to be plotted}

\item{...}{other arguments. Not used at the moment}
}
\description{
produces a scatter plot of the expression matrix of a flowframe. Note that, it takes some time to display the plot.
}
