% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phytoFilterClass.R
\name{reducedFlowframe,MarginEvents-method}
\alias{reducedFlowframe,MarginEvents-method}
\title{accesor method for reduced flowframe (MarginEvent class)}
\usage{
\S4method{reducedFlowframe}{MarginEvents}(x)
}
\arguments{
\item{x}{an object of class MarginEvents}
}
\value{
reduced Flowframe method for MarginEvents
}
\description{
accesor method for reduced flowframe (MarginEvent class)
}
