% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retain.R
\name{retain}
\alias{retain}
\title{Decides if a file should be retiained or removed based on its status
(from the good_measurement() function) as well as the number of particles
it contains}
\usage{
retain(x, y, dil_x = "100", dil_y = "200")
}
\arguments{
\item{x}{dataframe result from good_measurement() function}

\item{y}{dataframe result from good_measurement() function}

\item{dil_x}{dilution level of x}

\item{dil_y}{dilution level of y}
}
\value{
dataframe with status, dilution, filename of the retained file (either x or y)
}
\description{
Decides if a file should be retiained or removed based on its status
(from the good_measurement() function) as well as the number of particles
it contains
}
\examples{

\dontrun{
  retain(goodFiles_200, goodFiles_500, "200", "500")
}

}
